part the bl

+++ dotNet2022_8090_7731/BL/BL/DroneToList.cs
@@ -12,7 +12,7 @@ namespace IBL.BO
         public int Id { get; set; }
         public int Model { get; set; }
         public WeightCategories Weight { get; set; }
-        public float BatteryStatus { get; set; }
+        public double BatteryStatus { get; set; }
         public DroneStatus DStatus { get; set; }
         public Location CurrLocation { get; set; }
         public int NumOfParcel{ get; set; }
